Fix: Exported providers can start without packageName or leading dot
The existing exported content providers detection mechanism assumes that
an entry in the AndroidManifest.xml lists the full provider class name
or starts with a dot (".") followed by its class name.

However, entries can list only the class name of a provider without a
leading dot. Soot needs to know the full class name in order to resolve
it correctly.